Training_and_processing_videos
Not a great stopping point, "from_thermal_videos_to_tracks" is not working. It appears that FLIR SDK and Detectron do not play nicely with each other. But I am training the model with all annotations now and will take a break before coming back to process the images for predictions